app/requirements/base.txt
Django==3.2
psycopg2-binary==2.9.3
Pillow==9.0.0
django-markdownx==3.0.1
Markdown==3.3.6
stripe==2.65.0
django-lifecycle==0.9.3
django-allauth==0.47.0
factory_boy==3.2.1
django-cors-headers==3.11.0
whitenoise[brotli]==5.3.0
dj-database-url==0.5.0
boto3==1.20.40
django-storages==1.12.2
scout-apm==2.23.5
django-redis==5.2.0
django-clearcache==1.2.1
django-embed-video==1.4.5
djangorestframework==3.14.0